MIP in 'add_feed' command
This currently passes the test suite, but that reflects that the test
suite is woefully incomplete more than anything.

This changeset adds some new deps; primarily HTML::StripScripts::Parser.

Still to be done: test that the item addition code actually does what
it's supposed to -- that requires actually adding some items to the
test RSS feed.

Should also break down the tests a bit more and aim for more unit
testing of the component methods in addition to the end to end tests.